Rename common package
