+++ page/index.js
@@ -1,6 +1,7 @@
-//Dependency
-var React = require('react');
-var detailMixin = require('./detail').mixin;
+// Dependencies
+
+let React = require('react');
+let detailMixin = require('./detail').mixin;
 
 //Function to help page creation.
 module.exports = {
@@ -8,12 +9,12 @@ module.exports = {
   search: require('./search'),
   /**
    * Helper to creates a detail page.
-   * @param {object} config - The page configtration.
+   * @param {object} config - The page configuration.
    * @returns {object} - The react component associated to the page.
    */
   createDetail: function createDetail(config){
     config = config || {};
-    if(config.mixins !== undefined){
+    if (config.mixins) {
       config.mixins.push(detailMixin);
     } else {
       config.mixins = [detailMixin];
