+++ page/search/common/scroll-info-mixin.js
