+++ page/search/common/single-group-component.js
