+++ page/search/quick-search/index.js
@@ -13,9 +13,9 @@ let List = require('../../../list/selection').list.component;
 
 // Mixins
 
-let ScrollInfoMixin = require('../common-mixin/scroll-info-mixin').mixin;
-let GroupByMixin= require('../common-mixin/group-by-mixin').mixin;
-let SearchMixin = require('../common-mixin/search-mixin').mixin;
+let ScrollInfoMixin = require('../common/scroll-info-mixin').mixin;
+let GroupByMixin= require('../common/group-by-mixin').mixin;
+let SearchMixin = require('../common/search-mixin').mixin;
 
 /**
  * General search mixin.
